Simplify PolyIn design
- Conform CoreData stack to PolyIn protocol.
